requirements.txt
asgiref==3.5.0
astroid==2.11.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.4
Django==4.0.4
django-model-utils==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flake8==4.0.1
idna==3.3
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
openapi-codec==1.3.2
pep8==1.7.1
platformdirs==2.5.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
pylint==2.13.7
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
